safari-600.7-branch/LayoutTests/platform/wk2/http/tests/loading/basic-credentials-sent-automatically-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
didReceiveAuthenticationChallenge - Responding with first:first-pw
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
didReceiveAuthenticationChallenge - Responding with second:second-pw
frame "<!--framePath //<!--frame1-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame2-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame3-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
This test makes sure that once an HTTP Basic Auth. protected path is authenticated once, urls that emanate from that path automatically have their credentials sent without a challenge.
The first frame's path is /loading/resources/subresources/protected-resource.php, and we should get a challenge for it.
It will be authorized with first/first-pw.
The second frame's path is /loading/resources/protected-resource.php, and we should get a challenge for it, because it does not share a common subpath of the previously authorized resource.
It will be authorized with second/second-pw.
The third frame's path is also /loading/resources/protected-resource.php, and we should *not* get a challenge for it because we authorized to this path for the second frame.
It will be authorized with second/second-pw.
The fourth frame's path is /loading/resources/othersubresources/protected-resource.php, and we should *not* get a challenge for it, because it has a common subpath with the previously authorized second and third frames.
It will be authorized with second/second-pw.
  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Authenticated as user: first password: first-pw

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Authenticated as user: second password: second-pw

--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
Authenticated as user: second password: second-pw

--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
Authenticated as user: second password: second-pw
